Split PeerProtocol.scala into multiple files in newly created peer.protocol package
